[7544] - refactor csv import
