+++ app/models/lead_partner.rb
@@ -63,4 +63,15 @@ class LeadPartner < ApplicationRecord
   def name_and_code
     name
   end
+
+  def find_by_ukprn_or_urn(str)
+    return if str.blank?
+
+    case str.length
+    when 8
+      LeadPartner.find_by(ukprn: str)
+    when 6
+      LeadPartner.find_by(urn: str)
+    end
+  end
 end
